api_app/requirements.txt
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1